+++ src/nd2tool.c
@@ -3,7 +3,7 @@
 
 /* The stripped version of the header file from www.nd2sdk.com
  * where comments are removed.
-*/
+ */
 #include "Nd2ReadSdk_stripped.h"
 
 #include "tiff_util.h"
@@ -23,6 +23,11 @@ typedef struct{
     int showcoords;
     int overwrite;
     int composite;
+
+    /* One file per z-plane according to SpaceTx,
+       see https://github.com/elgw/nd2tool/issues/4 */
+    int save_individual_planes;
+
     nt_purpose purpose;
     char * fov_string; /* Specifying what fov to use */
     int meta_file;
@@ -97,7 +102,7 @@ typedef struct
 
 /*
  * Forward declarations
-*/
+ */
 
 static ntconf_t * ntconf_new(void);
 static void ntconf_free(ntconf_t * );
@@ -663,7 +668,7 @@ static void ensure_output_folder(ntconf_t * conf, nd2info_t * info)
 }
 
 /** @brief Write an ND2 file as one file per FOV and channel */
-static void nd2_to_tiff_split(void * nd2, ntconf_t * conf, nd2info_t * info)
+static void nd2_to_tiff_splitC(void * nd2, ntconf_t * conf, nd2info_t * info)
 {
 
     /* Prepare metadata for the tiff files */
@@ -805,10 +810,165 @@ static void nd2_to_tiff_split(void * nd2, ntconf_t * conf, nd2info_t * info)
     ttags_free(&tags);
 }
 
+
+/** @brief Write an ND2 file as one file per FOV and channel */
+static void nd2_to_tiff_splitC_splitZ(void * nd2, ntconf_t * conf, nd2info_t * info)
+{
+
+    /* Prepare metadata for the tiff files */
+    ttags * tags = ttags_new();
+    {
+        char * sw_string = ckcalloc(1024, 1);
+        sprintf(sw_string, "github.com/elgw/nd2tool source image: %s",
+                info->filename);
+        ttags_set_software(tags , sw_string);
+        free(sw_string);
+    }
+
+    int nchan = info->meta_att->nchannels;
+    int M = info->meta_att->channels[0]->M;
+    int N = info->meta_att->channels[0]->N;
+    int P = info->meta_att->channels[0]->P;
+
+    ttags_set_imagesize(tags, M, N, 1);
+    ttags_set_pixelsize_nm(tags,
+                           info->meta_att->channels[0]->dx_nm,
+                           info->meta_att->channels[0]->dy_nm,
+                           info->meta_att->channels[0]->dz_nm);
+
+
+    /* We choose to extract the image data multiple times and only
+     * collect pixels from one channel at a time. This is of course
+     * slightly slower than extracting all channels for a given FOV at
+     * a time but gives more predictable memory usage. */
+
+    /* Buffer for one slice and one color */
+    uint16_t * S = ckcalloc(M*N, sizeof(uint16_t));
+    LIMPICTURE * pic = ckcalloc(1, sizeof(LIMPICTURE));
+    Lim_InitPicture(pic, M, N, 16, nchan);
+
+    for(int64_t ff = 0; ff<info->nFOV; ff++) /* For each FOV */
+    {
+        if(conf->fov_string != NULL)
+        {
+            if(atoi(conf->fov_string) != (ff+1))
+            {
+                continue;
+            }
+        }
+
+        for(int64_t cc = 0; cc<nchan; cc++) /* For each channel */
+        {
+            for(int64_t kk = 0; kk<P; kk++) /* For each plane */
+            {
+
+                /* Write out to disk */
+                char * outname = ckcalloc(1024, 1);
+                /* SpaceTx
+                 * <image_type>-f<fov_id>-r<round_label>-c<ch_label>-z<zplane_label>.
+                 * Example: nuclei-f0-r2-c3-z33.tiff
+                 */
+                sprintf(outname, "%s/%s_f%ld-r%d-c%ld-z%lu.tif",
+                        info->outfolder,
+                        info->outfolder, /* <image_type> */
+                        ff, /* <fov_id> */
+                        0, /* <round_label> */
+                        cc, /* <ch_label> */
+                        kk); /* <zplane_label> */
+
+                printf("%s ", outname);
+                nd2info_log(info, "%s ", outname);
+
+                if(conf->shake)
+                {
+                    check_stage_position(info, ff, cc);
+                }
+
+                if(conf->overwrite == 0)
+                {
+                    if(isfile(outname))
+                    {
+                        printf("-- skipping, file exists\n");
+                        nd2info_log(info, "-- skipping, file exists\n");
+
+                        goto next_file;
+                    }
+                }
+                if(conf->verbose > 0)
+                {
+                    printf("... writing ... "); fflush(stdout);
+                }
+
+                if(conf->dry)
+                {
+                    printf(" (--dry, not writing)\n");
+                    goto next_file;
+                }
+
+                /* Create temporary file */
+                char * outname_tmp = ckcalloc(strlen(outname) + 16, 1);
+
+                sprintf(outname_tmp, "%s_tmp_XXXXXX", outname);
+                int tfid = 0;
+                if((tfid = mkstemp(outname_tmp)) == -1)
+                {
+                    fprintf(stderr, "Failed to create a temporary file based on pattern: %s\n", outname_tmp);
+                    exit(EXIT_FAILURE);
+                }
+                close(tfid);
+
+                tiff_writer_t * tw = tiff_writer_init(outname_tmp, tags, M, N, 1);
+
+                /* Returns interlaced data */
+                int res = Lim_FileGetImageData(nd2,
+                                               kk + ff*P, //uiSeqIndex,
+                                               pic);
+                if(res != 0)
+                {
+                    fprintf(stderr, "Failed to read from %s. At line %d\n",
+                            info->filename, __LINE__);
+                }
+
+                if( (pic->pImageData == NULL) || (pic->uiSize == 0) )
+                {
+                    fprintf(stderr, "Failed to retrieve image data\n");
+                }
+                uint16_t * pixels = (uint16_t *) pic->pImageData;
+
+                for(int64_t pp = 0; pp<M*N; pp++)
+                {
+                    S[pp] = pixels[pp*nchan+cc];
+                }
+                tiff_writer_write(tw, S);
+
+
+            /* Finish this image */
+            tiff_writer_finish(tw);
+            rename(outname_tmp, outname);
+            if(conf->verbose > 0)
+            {
+                printf("done\n");
+            }
+            nd2info_log(info, "\n");
+            free(outname_tmp);
+        next_file: ;
+            free(outname);
+            } // kk
+        } // cc
+    }// ff
+
+    Lim_DestroyPicture(pic);
+    free(pic);
+
+    free(S);
+    ttags_free(&tags);
+}
+
+
 /** @brief Write an ND2 file as composite tif files
  *
  * One file per FOV
-*/
+ */
 static void
 nd2_to_tiff_composite(void * nd2, ntconf_t * conf, nd2info_t * info)
 {
@@ -949,7 +1109,7 @@ nd2_to_tiff_composite(void * nd2, ntconf_t * conf, nd2info_t * info)
 /** @brief Try to convert an ND2 file to tif
  *
  * @return EXIT_SUCCESS or EXIT_FAILURE
-*/
+ */
 static int
 nd2_to_tiff(ntconf_t * conf, nd2info_t * info)
 {
@@ -1004,7 +1164,12 @@ nd2_to_tiff(ntconf_t * conf, nd2info_t * info)
     {
         nd2_to_tiff_composite(nd2, conf, info);
     } else {
-        nd2_to_tiff_split(nd2, conf, info);
+        if(conf->save_individual_planes)
+        {
+            nd2_to_tiff_splitC_splitZ(nd2, conf, info);
+        } else {
+            nd2_to_tiff_splitC(nd2, conf, info);
+        }
     }
 
     Lim_FileClose(nd2);
@@ -1192,6 +1357,8 @@ static void show_help(char * name)
            "for each file, generate a script to run deconwolf\n");
     printf("  --dry\n\t"
            "Perform a dry run, i.e. do not write files or create folders\n");
+    printf("  --SpaceTx\n\t"
+           "Save one image per z-plane according to the SpaceTx convention.");
     printf("Raw meta data extraction to stdout:\n");
     printf("  --meta\n\t all metadata.\n");
     printf("  --meta-file\n\t Lim_FileGetMetadata JSON.\n");
@@ -1239,6 +1406,7 @@ static int argparse(ntconf_t * conf, int argc, char ** argv)
         { "info",       no_argument, NULL, 'i'},
         { "overwrite",  no_argument, NULL, 'o'},
         { "shake",      no_argument, NULL, 's'},
+        { "SpaceTx",    no_argument, NULL, 'S'},
         { "test",       no_argument, NULL, 't'},
         { "verbose",    required_argument, NULL, 'v'},
         { "version",    no_argument, NULL, 'V'},
@@ -1253,7 +1421,7 @@ static int argparse(ntconf_t * conf, int argc, char ** argv)
     };
     int ch;
 
-    while((ch = getopt_long(argc, argv, "123456Fcdhiosv:CDVt",
+    while((ch = getopt_long(argc, argv, "123456FcdhiosSv:CDVt",
                             longopts, NULL)) != -1)
     {
         switch(ch) {
@@ -1312,6 +1480,9 @@ static int argparse(ntconf_t * conf, int argc, char ** argv)
             break;
         case 's':
             conf->shake = 1;
+            break;
+        case 'S':
+            conf->save_individual_planes = 1;
             break;
         case 't':
             nd2tool_util_ut();
@@ -1504,7 +1675,7 @@ static void showmeta(ntconf_t * conf, char * file)
 
 
 static void hello_log(__attribute__((unused)) ntconf_t * conf,
-               nd2info_t * info, int argc, char ** argv)
+                      nd2info_t * info, int argc, char ** argv)
 {
     if(conf->dry)
     {